improved testing framework
