+++ examples/test.js
@@ -1,28 +1,9 @@
 const hello = require("./index");
+const { test, expect } = require("./testing-framework");
 
-function test(title, callback) {
-  try {
-    callback();
-  } catch (error) {
-    console.log("Failed " + title + "...");
-    throw new Error(error);
-  }
-}
-
-function expect(result) {
-  return {
-    toBe(expected) {
-      if (result !== expected) {
-        throw new Error(`expected ${expected} but result is ${result}`);
-      } else {
-        console.log("Passed");
-      }
-    },
-  };
-}
-
-test("Test", () => {
+test("Test expected message", () => {
   const result = hello();
   const expected = "Hello World!";
   expect(result).toBe(expected);
+  expect(typeof result).toBe("string");
 });
